+++ config/src/parameters/user/boilerplate.rs
@@ -190,7 +190,7 @@ impl FromEnv for RootPartial {
 pub struct GenesisPartial {
     pub public_key: UserField<PublicKey>,
     pub file: UserField<PathBuf>,
-    pub encoded_config: UserField<String>,
+    pub signature: UserField<String>,
 }
 
 impl UnwrapPartial for GenesisPartial {
@@ -204,12 +204,12 @@ impl UnwrapPartial for GenesisPartial {
 
         let file = self.file.get();
 
-        let encoded_config = self.encoded_config.get();
+        let signature = self.signature.get();
 
         Ok(Genesis {
             public_key,
             file,
-            encoded_config,
+            signature,
         })
     }
 }
@@ -232,11 +232,11 @@ impl FromEnv for GenesisPartial {
         let file =
             ParseEnvResult::parse_simple(&mut emitter, env, "GENESIS_FILE", "genesis.file").into();
 
-        let encoded_config = ParseEnvResult::parse_simple(
+        let signature = ParseEnvResult::parse_simple(
             &mut emitter,
             env,
-            "ENCODED_CONFIG",
-            "genesis.encoded_config",
+            "SIGNATURE",
+            "genesis.signature",
         )
         .into();
 
@@ -245,7 +245,7 @@ impl FromEnv for GenesisPartial {
         Ok(Self {
             public_key,
             file,
-            encoded_config,
+            signature,
         })
     }
 }
