+++ config/tests/fixtures/minimal_alone_with_genesis.toml
@@ -3,4 +3,4 @@ extends = "base.toml"
 [genesis]
 file = "./empty_ok_genesis.json"
 public_key = "ed01208BA62848CF767D72E7F7F4B9D2D7BA07FEE33760F79ABE5597A51520E292A0CB"
-encoded_config = "9030303030303030302d303030302d303030302d303030302d303030303030303030303030ea9933ab8e0100000400804c390aef5d919d5506756ac51f7dd7e768da13fa60413fd5dc136851e9117213010145dd0ec90d6982851cdac085b54085a0e78689f14844742dcfc7da15b6ef78055584658cad5b9d69da200fa3997e490bfea2d7c417be9f6221d5ca89ba650008"
+signature = "9030303030303030302d303030302d303030302d303030302d303030303030303030303030ea9933ab8e0100000400804c390aef5d919d5506756ac51f7dd7e768da13fa60413fd5dc136851e9117213010145dd0ec90d6982851cdac085b54085a0e78689f14844742dcfc7da15b6ef78055584658cad5b9d69da200fa3997e490bfea2d7c417be9f6221d5ca89ba650008"
